backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
groq==0.4.0
tavily-python==0.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
chromadb==0.4.18
pypdf==3.17.1
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.25.2